Moved `optimizers` to `dion` to `from dion import Dion, DionParamConfig`
