+++ dion/scalar_opts.py
