+++ train.py
@@ -22,11 +22,11 @@ from typing import Optional
 
 from models.gpt_model import GPT, GPTConfig, parallelize_gpt_model
 from models.gpt_utils import DistributedDataLoader
-from optimizers.dion import Dion, DionMixedPrecisionConfig
-from optimizers.dion_reference import Dion as DionReference
-from optimizers.dion_simple import Dion as DionSimple
-from optimizers.muon import Muon
-from optimizers.muon_reference import Muon as MuonReference
+from dion import Dion, DionMixedPrecisionConfig
+from dion import DionReference
+from dion import DionSimple
+from dion import Muon
+from dion import MuonReference
 
 
 @dataclass
